Add PO sequnce nuber service.
This servise generates numbers in the fomat PO-2025-000123.
which is easy for business usecases, This is just based on assumptions.

Mechanism used here is SQL sequence objects. In real world distributed systems
Redis can be used for these type of number genrations.

I wanted to showcase this as a demostration, insted of using auto increment ids.